refactor: extract research universe from runner
