+++ research/run_research.py
@@ -4,25 +4,10 @@ voert alle enabled strategieën uit via de registry
 en schrijft resultaten naar CSV + latest JSON.
 """
 
-from datetime import datetime, timedelta, UTC
-
 from agent.backtesting.engine import BacktestEngine
 from research.registry import get_enabled_strategies
 from research.results import make_result_row, append_results_to_csv, write_latest_json
-
-
-ASSETS = ["BTC-USD", "ETH-USD"]
-INTERVALS = ["1h", "4h"]
-
-def get_date_range(interval):
-    now = datetime.now(UTC)
-
-    if interval in ["1h", "4h"]:
-        start = now - timedelta(days=700)
-    else:
-        start = now - timedelta(days=1500)
-
-    return start.strftime("%Y-%m-%d"), now.strftime("%Y-%m-%d")
+from research.universe import ASSETS, INTERVALS, get_date_range
 
 def run_research():
     rows = []
@@ -41,7 +26,7 @@ def run_research():
                     metrics = engine.grid_search(
                         strategie_factory=strategy["factory"],
                         param_grid=strategy["params"],
-                        assets=[asset],
+                        assets=[asset.symbol],
                         interval=interval,
                     )
 
@@ -49,7 +34,7 @@ def run_research():
 
                     row = make_result_row(
                         strategy=strategy,
-                        asset=asset,
+                        asset=asset.symbol,
                         interval=interval,
                         params=params_used,
                         metrics=metrics,
@@ -57,7 +42,7 @@ def run_research():
                 except Exception as e:
                     row = make_result_row(
                         strategy=strategy,
-                        asset=asset,
+                        asset=asset.symbol,
                         interval=interval,
                         params={},
                         metrics={},
